There are people who think that things that happen in fiction do not really happen. These people are wrong. (v0.4.0) (#10)
This PR states the largest update to fiction yet. Its architecture has been completely reworked to create a platform upon which the FCN community can build. To this end, fiction provides (1) a header-only template library that can be utilized in custom projects, (2) a framework for experiments for rapid prototyping and reproducible findings, and (3) the established CLI that is now built with the library.

Furthermore, the build system has been revised and third-party dependencies have been cut to leave behind only what was really necessary. No more Boost means less trouble compiling on any system. Z3 is now an optional dependency that, if installed, enables additional features.

Finally, there is a whole lot of new features like support for the Silicon Dangling Bond (SiDB) technology, SiQAD and QCA-STACK file formats, the integration of a technology mapper, layout types with hexagonal tiles, and much more.